ValidatorBoxFlyweight defined
